Measure the completion time to send a fixed amount of data
Added a new flag "-totalFlow" to measure the total completion
time takes to send a fixed amount of data over a connection.
The value for "-totalFlow" is an uint64. When the "-totalFlow"
is on, the timeout feature ("-totalDuration") is disabled. The
program will automatically exit when all the data are sent.
To enable the "-totalFlow", server side needs to provide two
extra flags "-totalFlow" and "-passiveServer". The values of
"-totalFlow" for both client side and server side should be
the same.

 Changes to be committed:
	modified:   client.go
	modified:   csv.go
	modified:   main.go
	modified:   server.go